+++ src/Networking.h
@@ -14,9 +14,10 @@ class NetworkingInterface
 public:
     virtual ~NetworkingInterface() = default;
 
-    virtual void sendToClient(int toClientID, const Message &message) = 0;
-    virtual std::vector<std::pair<int, Message> > receiveFromClients() = 0;
-    virtual std::vector<int> getConnectedClientIDs() const = 0;
+    virtual void sendToClient(uintptr_t toClientID, const Message &message) = 0;
+    // virtual void sendMessageToServer(int toClientID, const Message &message) = 0;
+    virtual std::vector<std::pair<uintptr_t, Message> > receiveFromClients() = 0;
+    virtual std::vector<uintptr_t> getConnectedClientIDs() const = 0;
 };
 
 // TODO (if interface makes sense):
